+++ packages/std/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.17.0] - 2026-08-12
+
 ### Changed
 
 - Updated OTF Web framework dependencies (`@opentf/web` to `^0.27.0`, `@opentf/web-docs` to `^0.25.0`, `@opentf/web-cli` to `^1.25.0`) and `@opentf/std` tarball dependency to `0.16.0`.